third_party/blink/web_tests/svg/dynamic-updates/SVGFEDisplacementMapElement-dom-scale-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'scale' attribute of the SVGFEDisplacementMapeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.getAttribute('scale') is "34"
PASS displacementMap.getAttribute('scale') is "64"
PASS successfullyParsed is true

TEST COMPLETE
